Keep the most recent record
Sometimes there will be multiple updates on an object in a repository,
resulting in more than one records with the same identifier between harvests.
Since oai-pmh response is not sorted by datestamp, we check the datestamp
before saving the record to make sure we are saving the most recent
version. This also handles deleted records.